requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
python-multipart>=0.0.6
pyjwt[crypto]>=2.8.0
passlib[bcrypt]>=1.7.4
opencv-python-headless>=4.9.0.80
numpy>=1.26.0
pillow>=10.2.0
reportlab>=4.0.9
pytesseract>=0.3.10
google-generativeai>=0.3.2
psycopg2-binary==2.9.9
openpyxl>=3.1.2
python-dotenv>=1.0.1
websockets>=12.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
aiofiles>=23.2.1
